+++ sdk/avs/arm-avs/src/models/index.ts
@@ -14,6 +14,10 @@ export type AddonPropertiesUnion =
   | AddonHcxProperties
   | AddonSrmProperties
   | AddonVrProperties;
+export type HostPropertiesUnion =
+  | HostProperties
+  | GeneralHostProperties
+  | SpecializedHostProperties;
 export type PlacementPolicyPropertiesUnion =
   | PlacementPolicyProperties
   | VmHostPlacementPolicyProperties
@@ -217,7 +221,7 @@ export interface IdentitySource {
   name?: string;
   /** The domain's NetBIOS name */
   alias?: string;
-  /** The domain's dns name */
+  /** The domain's DNS name */
   domain?: string;
   /** The base distinguished name for users */
   baseUserDN?: string;
@@ -399,6 +403,82 @@ export interface SystemData {
   lastModifiedAt?: Date;
 }
 
+/** Paged collection of ResourceSku items */
+export interface PagedResourceSku {
+  /** The ResourceSku items on this page */
+  value: ResourceSku[];
+  /** The link to the next page of items */
+  nextLink?: string;
+}
+
+/** A SKU for a resource. */
+export interface ResourceSku {
+  /** The type of resource the SKU applies to. */
+  resourceType: ResourceSkuResourceType;
+  /** The name of the SKU. */
+  name: string;
+  /** The tier of virtual machines in a scale set */
+  tier?: string;
+  /** The size of the SKU. */
+  size?: string;
+  /** The family of the SKU. */
+  family?: string;
+  /** The set of locations that the SKU is available. */
+  locations: string[];
+  /** A list of locations and availability zones in those locations where the SKU is available */
+  locationInfo: ResourceSkuLocationInfo[];
+  /** Name value pairs to describe the capability. */
+  capabilities?: ResourceSkuCapabilities[];
+  /** The restrictions of the SKU. */
+  restrictions: ResourceSkuRestrictions[];
+}
+
+/** Describes an available Compute SKU Location Information. */
+export interface ResourceSkuLocationInfo {
+  /** Location of the SKU */
+  location: string;
+  /** List of availability zones where the SKU is supported. */
+  zones: string[];
+  /** Gets details of capabilities available to a SKU in specific zones. */
+  zoneDetails: ResourceSkuZoneDetails[];
+}
+
+/** Describes The zonal capabilities of a SKU. */
+export interface ResourceSkuZoneDetails {
+  /** Gets the set of zones that the SKU is available in with the specified capabilities. */
+  name: string[];
+  /** A list of capabilities that are available for the SKU in the specified list of zones. */
+  capabilities: ResourceSkuCapabilities[];
+}
+
+/** Describes The SKU capabilities object. */
+export interface ResourceSkuCapabilities {
+  /** The name of the SKU capability. */
+  name: string;
+  /** The value of the SKU capability. */
+  value: string;
+}
+
+/** The restrictions of the SKU. */
+export interface ResourceSkuRestrictions {
+  /** the type of restrictions. */
+  type?: ResourceSkuRestrictionsType;
+  /** The value of restrictions. If the restriction type is set to location. This would be different locations where the SKU is restricted. */
+  values: string[];
+  /** The information about the restriction where the SKU cannot be used. */
+  restrictionInfo: ResourceSkuRestrictionInfo;
+  /** the reason for restriction. */
+  reasonCode?: ResourceSkuRestrictionsReasonCode;
+}
+
+/** Describes an available Compute SKU Restriction Information. */
+export interface ResourceSkuRestrictionInfo {
+  /** Locations where the SKU is restricted */
+  locations?: string[];
+  /** List of availability zones where the SKU is restricted. */
+  zones?: string[];
+}
+
 /** An update to a private cloud resource */
 export interface PrivateCloudUpdate {
   /** Resource tags. */
@@ -517,6 +597,49 @@ export interface DiskPoolVolume {
 export interface ElasticSanVolume {
   /** Azure resource ID of the Elastic SAN Volume */
   targetId: string;
+}
+
+/** A Pure Storage volume from PureStorage.Block provider */
+export interface PureStorageVolume {
+  /** Azure resource ID of the Pure Storage Pool */
+  storagePoolId: string;
+  /** Volume size to be used to create a Virtual Volumes (vVols) datastore */
+  sizeGb: number;
+}
+
+/** The response of a Host list operation. */
+export interface HostListResult {
+  /** The Host items on this page */
+  value: Host[];
+  /** The link to the next page of items */
+  nextLink?: string;
+}
+
+/** The properties of a host. */
+export interface HostProperties {
+  /** Polymorphic discriminator, which specifies the different types this object can be */
+  kind: "General" | "Specialized";
+  /**
+   * The state of the host provisioning.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly provisioningState?: HostProvisioningState;
+  /** Display name of the host in VMware vCenter. */
+  displayName?: string;
+  /**
+   * vCenter managed object reference ID of the host.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly moRefId?: string;
+  /**
+   * Fully qualified domain name of the host.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly fqdn?: string;
+  /** If provided, the host is in maintenance. The value is the reason for maintenance. */
+  maintenance?: HostMaintenance;
+  /** NOTE: This property will not be serialized. It can only be populated by the server. */
+  readonly faultDomain?: string;
 }
 
 /** List of all zones and associated hosts for a cluster */
@@ -638,6 +761,22 @@ export interface AdminCredentials {
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly vcenterPassword?: string;
+}
+
+/** The response of a ProvisionedNetwork list operation. */
+export interface ProvisionedNetworkListResult {
+  /** The ProvisionedNetwork items on this page */
+  value: ProvisionedNetwork[];
+  /** The link to the next page of items */
+  nextLink?: string;
+}
+
+/** The response of a PureStoragePolicy list operation. */
+export interface PureStoragePolicyListResult {
+  /** The PureStoragePolicy items on this page */
+  value: PureStoragePolicy[];
+  /** The link to the next page of items */
+  nextLink?: string;
 }
 
 /** The response of a ScriptExecution list operation. */
@@ -842,6 +981,10 @@ export interface AddonHcxProperties extends AddonProperties {
   addonType: "HCX";
   /** The HCX offer, example VMware MaaS Cloud Provider (Enterprise) */
   offer: string;
+  /** HCX management network. */
+  managementNetwork?: string;
+  /** HCX uplink network */
+  uplinkNetwork?: string;
 }
 
 /** The properties of a Site Recovery Manager (SRM) addon */
@@ -858,6 +1001,18 @@ export interface AddonVrProperties extends AddonProperties {
   addonType: "VR";
   /** The vSphere Replication Server (VRS) count */
   vrsCount: number;
+}
+
+/** The properties of a general host. */
+export interface GeneralHostProperties extends HostProperties {
+  /** Polymorphic discriminator, which specifies the different types this object can be */
+  kind: "General";
+}
+
+/** The properties of a specialized host. */
+export interface SpecializedHostProperties extends HostProperties {
+  /** Polymorphic discriminator, which specifies the different types this object can be */
+  kind: "Specialized";
 }
 
 /** VM-Host placement policy properties */
@@ -946,6 +1101,8 @@ export interface PrivateCloud extends TrackedResource {
   sku: Sku;
   /** The managed service identities assigned to this resource. */
   identity?: PrivateCloudIdentity;
+  /** The availability zones. */
+  zones?: string[];
   /** The default cluster used for management */
   managementCluster?: ManagementCluster;
   /** Connectivity to internet is enabled or disabled */
@@ -1116,11 +1273,44 @@ export interface Datastore extends ProxyResource {
   diskPoolVolume?: DiskPoolVolume;
   /** An Elastic SAN volume */
   elasticSanVolume?: ElasticSanVolume;
+  /** A Pure Storage volume */
+  pureStorageVolume?: PureStorageVolume;
   /**
    * The operational status of the datastore
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly status?: DatastoreStatus;
+}
+
+/** A host resource */
+export interface Host extends ProxyResource {
+  /** The availability zones. */
+  zones?: string[];
+  /** The SKU (Stock Keeping Unit) assigned to this resource. */
+  sku?: Sku;
+  /** The kind of host */
+  kind?: HostKind;
+  /**
+   * The state of the host provisioning.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly provisioningState?: HostProvisioningState;
+  /** Display name of the host in VMware vCenter. */
+  displayName?: string;
+  /**
+   * vCenter managed object reference ID of the host.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly moRefId?: string;
+  /**
+   * Fully qualified domain name of the host.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly fqdn?: string;
+  /** If provided, the host is in maintenance. The value is the reason for maintenance. */
+  maintenance?: HostMaintenance;
+  /** NOTE: This property will not be serialized. It can only be populated by the server. */
+  readonly faultDomain?: string;
 }
 
 /** A vSphere Distributed Resource Scheduler (DRS) placement policy */
@@ -1142,7 +1332,7 @@ export interface VirtualMachine extends ProxyResource {
    */
   readonly displayName?: string;
   /**
-   * Virtual machine managed object reference id
+   * vCenter managed object reference ID of the virtual machine
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly moRefId?: string;
@@ -1221,6 +1411,38 @@ export interface IscsiPath extends ProxyResource {
   readonly provisioningState?: IscsiPathProvisioningState;
   /** CIDR Block for iSCSI path. */
   networkBlock?: string;
+}
+
+/** A provisioned network resource */
+export interface ProvisionedNetwork extends ProxyResource {
+  /**
+   * The provisioning state of the resource.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly provisioningState?: ProvisionedNetworkProvisioningState;
+  /**
+   * The address prefixes of the provisioned network in CIDR notation.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly addressPrefix?: string;
+  /**
+   * The type of network provisioned.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly networkType?: ProvisionedNetworkTypes;
+}
+
+/** An instance describing a Pure Storage Policy Based Management policy */
+export interface PureStoragePolicy extends ProxyResource {
+  /** Definition of a Pure Storage Policy Based Management policy */
+  storagePolicyDefinition?: string;
+  /** Azure resource ID of the Pure Storage Pool associated with the storage policy */
+  storagePoolId?: string;
+  /**
+   * The state of the Pure Storage Policy Based Management policy provisioning
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly provisioningState?: PureStoragePolicyProvisioningState;
 }
 
 /** An instance of a script executed by a user - custom or AVS */
@@ -1693,6 +1915,22 @@ export interface IscsiPathsCreateOrUpdateHeaders {
 
 /** Defines headers for IscsiPaths_delete operation. */
 export interface IscsiPathsDeleteHeaders {
+  /** The Location header contains the URL where the status of the long running operation can be checked. */
+  location?: string;
+  /** The Retry-After header can indicate how long the client should wait before polling the operation status. */
+  retryAfter?: number;
+}
+
+/** Defines headers for PureStoragePolicies_createOrUpdate operation. */
+export interface PureStoragePoliciesCreateOrUpdateHeaders {
+  /** A link to the status monitor */
+  azureAsyncOperation?: string;
+  /** The Retry-After header can indicate how long the client should wait before polling the operation status. */
+  retryAfter?: number;
+}
+
+/** Defines headers for PureStoragePolicies_delete operation. */
+export interface PureStoragePoliciesDeleteHeaders {
   /** The Location header contains the URL where the status of the long running operation can be checked. */
   location?: string;
   /** The Retry-After header can indicate how long the client should wait before polling the operation status. */
@@ -2186,6 +2424,60 @@ export enum KnownCreatedByType {
  */
 export type CreatedByType = string;
 
+/** Known values of {@link ResourceSkuResourceType} that the service accepts. */
+export enum KnownResourceSkuResourceType {
+  /** The SKU is for a private cloud. */
+  PrivateClouds = "privateClouds",
+  /** The SKU is for a private cloud cluster. */
+  PrivateCloudsClusters = "privateClouds/clusters",
+}
+
+/**
+ * Defines values for ResourceSkuResourceType. \
+ * {@link KnownResourceSkuResourceType} can be used interchangeably with ResourceSkuResourceType,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **privateClouds**: The SKU is for a private cloud. \
+ * **privateClouds\/clusters**: The SKU is for a private cloud cluster.
+ */
+export type ResourceSkuResourceType = string;
+
+/** Known values of {@link ResourceSkuRestrictionsType} that the service accepts. */
+export enum KnownResourceSkuRestrictionsType {
+  /** SKU restricted by location. */
+  Location = "Location",
+  /** SKU restricted by availability zone. */
+  Zone = "Zone",
+}
+
+/**
+ * Defines values for ResourceSkuRestrictionsType. \
+ * {@link KnownResourceSkuRestrictionsType} can be used interchangeably with ResourceSkuRestrictionsType,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Location**: SKU restricted by location. \
+ * **Zone**: SKU restricted by availability zone.
+ */
+export type ResourceSkuRestrictionsType = string;
+
+/** Known values of {@link ResourceSkuRestrictionsReasonCode} that the service accepts. */
+export enum KnownResourceSkuRestrictionsReasonCode {
+  /** The restriction is due to exceeding a quota limitation. */
+  QuotaId = "QuotaId",
+  /** The restriction is not available for this subscription. */
+  NotAvailableForSubscription = "NotAvailableForSubscription",
+}
+
+/**
+ * Defines values for ResourceSkuRestrictionsReasonCode. \
+ * {@link KnownResourceSkuRestrictionsReasonCode} can be used interchangeably with ResourceSkuRestrictionsReasonCode,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **QuotaId**: The restriction is due to exceeding a quota limitation. \
+ * **NotAvailableForSubscription**: The restriction is not available for this subscription.
+ */
+export type ResourceSkuRestrictionsReasonCode = string;
+
 /** Known values of {@link AddonType} that the service accepts. */
 export enum KnownAddonType {
   /** SRM */
@@ -2401,6 +2693,63 @@ export enum KnownDatastoreStatus {
  * **DeadOrError**: is dead or error
  */
 export type DatastoreStatus = string;
+
+/** Known values of {@link HostKind} that the service accepts. */
+export enum KnownHostKind {
+  /** General */
+  General = "General",
+  /** Specialized */
+  Specialized = "Specialized",
+}
+
+/**
+ * Defines values for HostKind. \
+ * {@link KnownHostKind} can be used interchangeably with HostKind,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **General** \
+ * **Specialized**
+ */
+export type HostKind = string;
+
+/** Known values of {@link HostProvisioningState} that the service accepts. */
+export enum KnownHostProvisioningState {
+  /** Resource has been created. */
+  Succeeded = "Succeeded",
+  /** Resource creation failed. */
+  Failed = "Failed",
+  /** Resource creation was canceled. */
+  Canceled = "Canceled",
+}
+
+/**
+ * Defines values for HostProvisioningState. \
+ * {@link KnownHostProvisioningState} can be used interchangeably with HostProvisioningState,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Succeeded**: Resource has been created. \
+ * **Failed**: Resource creation failed. \
+ * **Canceled**: Resource creation was canceled.
+ */
+export type HostProvisioningState = string;
+
+/** Known values of {@link HostMaintenance} that the service accepts. */
+export enum KnownHostMaintenance {
+  /** The host is a replacement host. */
+  Replacement = "Replacement",
+  /** The host is for an upgrade, such as an upgrade to ESXi, NSX-T, or other component. */
+  Upgrade = "Upgrade",
+}
+
+/**
+ * Defines values for HostMaintenance. \
+ * {@link KnownHostMaintenance} can be used interchangeably with HostMaintenance,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Replacement**: The host is a replacement host. \
+ * **Upgrade**: The host is for an upgrade, such as an upgrade to ESXi, NSX-T, or other component.
+ */
+export type HostMaintenance = string;
 
 /** Known values of {@link PlacementPolicyType} that the service accepts. */
 export enum KnownPlacementPolicyType {
@@ -2665,6 +3014,87 @@ export enum KnownIscsiPathProvisioningState {
  * **Updating**: is updating
  */
 export type IscsiPathProvisioningState = string;
+
+/** Known values of {@link ProvisionedNetworkProvisioningState} that the service accepts. */
+export enum KnownProvisionedNetworkProvisioningState {
+  /** Resource has been created. */
+  Succeeded = "Succeeded",
+  /** Resource creation failed. */
+  Failed = "Failed",
+  /** Resource creation was canceled. */
+  Canceled = "Canceled",
+}
+
+/**
+ * Defines values for ProvisionedNetworkProvisioningState. \
+ * {@link KnownProvisionedNetworkProvisioningState} can be used interchangeably with ProvisionedNetworkProvisioningState,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Succeeded**: Resource has been created. \
+ * **Failed**: Resource creation failed. \
+ * **Canceled**: Resource creation was canceled.
+ */
+export type ProvisionedNetworkProvisioningState = string;
+
+/** Known values of {@link ProvisionedNetworkTypes} that the service accepts. */
+export enum KnownProvisionedNetworkTypes {
+  /** network for ESX management */
+  EsxManagement = "esxManagement",
+  /** network for ESX replication */
+  EsxReplication = "esxReplication",
+  /** network for HCX management */
+  HcxManagement = "hcxManagement",
+  /** network for HCX uplink */
+  HcxUplink = "hcxUplink",
+  /** network for vCenter management */
+  VcenterManagement = "vcenterManagement",
+  /** network for vmotion */
+  Vmotion = "vmotion",
+  /** network for vsan */
+  Vsan = "vsan",
+}
+
+/**
+ * Defines values for ProvisionedNetworkTypes. \
+ * {@link KnownProvisionedNetworkTypes} can be used interchangeably with ProvisionedNetworkTypes,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **esxManagement**: network for ESX management \
+ * **esxReplication**: network for ESX replication \
+ * **hcxManagement**: network for HCX management \
+ * **hcxUplink**: network for HCX uplink \
+ * **vcenterManagement**: network for vCenter management \
+ * **vmotion**: network for vmotion \
+ * **vsan**: network for vsan
+ */
+export type ProvisionedNetworkTypes = string;
+
+/** Known values of {@link PureStoragePolicyProvisioningState} that the service accepts. */
+export enum KnownPureStoragePolicyProvisioningState {
+  /** Resource has been created. */
+  Succeeded = "Succeeded",
+  /** Resource creation failed. */
+  Failed = "Failed",
+  /** Resource creation was canceled. */
+  Canceled = "Canceled",
+  /** is deleting */
+  Deleting = "Deleting",
+  /** is updating */
+  Updating = "Updating",
+}
+
+/**
+ * Defines values for PureStoragePolicyProvisioningState. \
+ * {@link KnownPureStoragePolicyProvisioningState} can be used interchangeably with PureStoragePolicyProvisioningState,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Succeeded**: Resource has been created. \
+ * **Failed**: Resource creation failed. \
+ * **Canceled**: Resource creation was canceled. \
+ * **Deleting**: is deleting \
+ * **Updating**: is updating
+ */
+export type PureStoragePolicyProvisioningState = string;
 
 /** Known values of {@link ScriptExecutionParameterType} that the service accepts. */
 export enum KnownScriptExecutionParameterType {
@@ -3425,6 +3855,19 @@ export interface PrivateCloudsListNextOptionalParams
 export type PrivateCloudsListNextResponse = PrivateCloudList;
 
 /** Optional parameters. */
+export interface SkusListOptionalParams extends coreClient.OperationOptions {}
+
+/** Contains response data for the list operation. */
+export type SkusListResponse = PagedResourceSku;
+
+/** Optional parameters. */
+export interface SkusListNextOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the listNext operation. */
+export type SkusListNextResponse = PagedResourceSku;
+
+/** Optional parameters. */
 export interface AddonsListOptionalParams extends coreClient.OperationOptions {}
 
 /** Contains response data for the list operation. */
@@ -3652,6 +4095,25 @@ export interface DatastoresListNextOptionalParams
 export type DatastoresListNextResponse = DatastoreList;
 
 /** Optional parameters. */
+export interface HostsListOptionalParams extends coreClient.OperationOptions {}
+
+/** Contains response data for the list operation. */
+export type HostsListResponse = HostListResult;
+
+/** Optional parameters. */
+export interface HostsGetOptionalParams extends coreClient.OperationOptions {}
+
+/** Contains response data for the get operation. */
+export type HostsGetResponse = Host;
+
+/** Optional parameters. */
+export interface HostsListNextOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the listNext operation. */
+export type HostsListNextResponse = HostListResult;
+
+/** Optional parameters. */
 export interface PlacementPoliciesListOptionalParams
   extends coreClient.OperationOptions {}
 
@@ -3855,6 +4317,73 @@ export interface IscsiPathsListByPrivateCloudNextOptionalParams
 
 /** Contains response data for the listByPrivateCloudNext operation. */
 export type IscsiPathsListByPrivateCloudNextResponse = IscsiPathListResult;
+
+/** Optional parameters. */
+export interface ProvisionedNetworksListOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the list operation. */
+export type ProvisionedNetworksListResponse = ProvisionedNetworkListResult;
+
+/** Optional parameters. */
+export interface ProvisionedNetworksGetOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the get operation. */
+export type ProvisionedNetworksGetResponse = ProvisionedNetwork;
+
+/** Optional parameters. */
+export interface ProvisionedNetworksListNextOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the listNext operation. */
+export type ProvisionedNetworksListNextResponse = ProvisionedNetworkListResult;
+
+/** Optional parameters. */
+export interface PureStoragePoliciesListOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the list operation. */
+export type PureStoragePoliciesListResponse = PureStoragePolicyListResult;
+
+/** Optional parameters. */
+export interface PureStoragePoliciesGetOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the get operation. */
+export type PureStoragePoliciesGetResponse = PureStoragePolicy;
+
+/** Optional parameters. */
+export interface PureStoragePoliciesCreateOrUpdateOptionalParams
+  extends coreClient.OperationOptions {
+  /** Delay to wait until next poll, in milliseconds. */
+  updateIntervalInMs?: number;
+  /** A serialized poller which can be used to resume an existing paused Long-Running-Operation. */
+  resumeFrom?: string;
+}
+
+/** Contains response data for the createOrUpdate operation. */
+export type PureStoragePoliciesCreateOrUpdateResponse = PureStoragePolicy;
+
+/** Optional parameters. */
+export interface PureStoragePoliciesDeleteOptionalParams
+  extends coreClient.OperationOptions {
+  /** Delay to wait until next poll, in milliseconds. */
+  updateIntervalInMs?: number;
+  /** A serialized poller which can be used to resume an existing paused Long-Running-Operation. */
+  resumeFrom?: string;
+}
+
+/** Contains response data for the delete operation. */
+export type PureStoragePoliciesDeleteResponse =
+  PureStoragePoliciesDeleteHeaders;
+
+/** Optional parameters. */
+export interface PureStoragePoliciesListNextOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the listNext operation. */
+export type PureStoragePoliciesListNextResponse = PureStoragePolicyListResult;
 
 /** Optional parameters. */
 export interface ScriptExecutionsListOptionalParams